+++ source/renderer/zettlr-renderer.js
@@ -20,7 +20,6 @@ const ZettlrEditor = require('./zettlr-editor.js')
 const ZettlrBody = require('./zettlr-body.js')
 const ZettlrToolbar = require('./zettlr-toolbar.js')
 const ZettlrPomodoro = require('./zettlr-pomodoro.js')
-const popup = require('./zettlr-popup.js')
 const ZettlrAttachments = require('./zettlr-attachments.js')
 
 const { remote } = require('electron')
@@ -282,105 +281,6 @@ class ZettlrRenderer {
     */
   toggleAttachments () {
     this._attachments.toggle()
-  }
-
-  /**
-    * Displays a table of content.
-    * @return {void} (Point of) No return.
-    */
-  toc () {
-    if (this.getCurrentFile() === null) {
-      return
-    }
-
-    let toc = this._editor.buildTOC()
-
-    if (toc.length === 0) {
-      return
-    }
-
-    let idUniquifier = Date.now()
-
-    let cnt = $('<div id="toc-container-' + idUniquifier + '">')
-    let h1 = 0
-    let h2 = 0
-    let h3 = 0
-    let h4 = 0
-    let h5 = 0
-    let h6 = 0
-    for (let entry of toc) {
-      let level = ''
-      switch (entry.level) {
-        case 1:
-          h1++
-          h2 = h3 = h4 = h5 = h6 = 0
-          level = h1
-          break
-        case 2:
-          h2++
-          h3 = h4 = h5 = h6 = 0
-          level = [h1, h2].join('.')
-          break
-        case 3:
-          h3++
-          h4 = h5 = h6 = 0
-          level = [h1, h2, h3].join('.')
-          break
-        case 4:
-          h4++
-          h5 = h6 = 0
-          level = [h1, h2, h3, h4].join('.')
-          break
-        case 5:
-          h5++
-          h6 = 0
-          level = [h1, h2, h3, h4, h5].join('.')
-          break
-        case 6:
-          h6++
-          level = [h1, h2, h3, h4, h5, h6].join('.')
-      }
-
-      cnt.append(
-        $('<a>').text(level + '. ' + entry.text)
-          .attr('data-line', entry.line)
-          .attr('href', '#')
-          .addClass('toc-link')
-      )
-    }
-
-    let tocPopup = popup($('.button.show-toc'), cnt)
-
-    // On click jump to line
-    $('.toc-link').click((event) => {
-      let elem = $(event.target)
-      this._editor.jtl(elem.attr('data-line'))
-    })
-
-    // Sortable
-    $('#toc-container-' + idUniquifier).sortable({
-      axis: 'y',
-      items: '> .toc-link',
-      update: (event, ui) => {
-        // The user has dropped the item someplace else.
-        let newIndex = ui.item.index()
-        let originalLine = parseInt(ui.item.attr('data-line'))
-        let sumLength = $('#toc-container-' + idUniquifier + ' > .toc-link').length
-        if (newIndex < sumLength - 1) {
-          let elementBelow = $('#toc-container-' + idUniquifier + ' > .toc-link').eq(newIndex + 1)
-          let aboveLine = parseInt(elementBelow.attr('data-line'))
-          this._editor.moveSection(originalLine, aboveLine)
-        } else {
-          this._editor.moveSection(originalLine, -1)
-        }
-
-        // Cool, now destroy the sortable, rebuild the TOC, and re-fill the div
-        // again.
-        $('#toc-container-' + idUniquifier).sortable('destroy')
-        tocPopup.close()
-        this.toc()
-      }
-    })
   }
 
   /**